Performance improvements to ERTokenImpl.equals()
I noticed some performance impacts after my fix to equals() in ERTokenImpl, and some quality time with the profiler pointed me back to this method. The way I'd been doing the equals() check in ERTokenImpl relied on toString() and then a string comparison; toString() itself involved some expensive string concatenation, which was being doubled- once for the "source" token and once for the "other" token. Since equals() was getting used inside of a lot of inner loops, this means a lot of object allocation happening over and over, plus string comparison ain't free.

I have re-implemented equals() to directly compare the various fields; on the surface, this method might look more expensive than just doing toString() twice, since it involves three string comparisons in the worst case. However, the way the method is written, equals() begins with a simple int comparison that is likely to fail- the only time the string comparisons kick in is if the two tokens had the same offset, which is a relatively less common comparison. So we're getting a boost in the most common case. It's a little hard to say quantitatively how much this is getting us but _very_ anecdotally in my test input I'm seeing about a 5-10% speedup for long documents with a lot of small matches.