Moderate surgery on IVFLookup to enable term info caching.
There exists a class in MML called TermConceptInfoCache that looks like, at one point in time, it was meant for use as part of IVFLookup. However, as best as I can tell, it is not currently used anywhere- IVFLookup _makes_ an instance and keeps it around, but never actually _uses_ it for anything. Looking at it, a lot of its internal logic seems to duplicate things that are happening elsewhere in EntityLookup4/5. My best guess is that at some point in the past things were re-factored and the class's use was discontinued as a result.

This means that term info lookup is no longer being cached, and this (via the profiler) is definitely a hot spot in entity matching where caching has a major performance impact. I have put in an LRUCache here, and done some surgery on the class to support it use:

1. Adding a couple of new property keys, one to control whether caching is enabled and then another to set the LRUCache size.

2. Removal of the empty/blank constructor. This was actually being used in a couple of places, but in all of the contexts where it was being used, we had a properties object that we could have passed in, and so it seems accidental that it wasn't being used. So I went in to EntityLookup4/5 and changed how they are instantiating their IVFLookup objects to pass in their configuration properties.

At this point, I do think that g.n.n.n.m.lite.TermConceptInfoCache can probably be removed altogether but will defer to others on that decision.